Moved client-related elements from BSTListView to BSTClientInterface.
Files checked in: DataRepo/views/models/bst/client_interface.py DataRepo/views/models/bst/list_view.py DataRepo/views/utils.py
